% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_colors.R
\name{common_colors}
\alias{common_colors}
\title{Creates common colors}
\usage{
common_colors(vars, color_pal = "Set2")
}
\arguments{
\item{vars}{a vector containing all the categories}

\item{color_pal}{a character with name of color pallete from RColorBrewer}
}
\description{
Creates common colors
}
